extract AuthError in a separate file
