+++ lib/extensions.js
@@ -24,7 +24,8 @@ module.exports = function(realmConstructor) {
 
     // Add sync methods
     if (realmConstructor.Sync) {
-        Object.assign(realmConstructor.Sync, require('./sync'));
+        realmConstructor.Sync.User = require('./sync').User;
+        realmConstructor.Sync.AuthError = require('./errors').AuthError;
     }
 
     // TODO: Remove this now useless object.
