+++ lib/sync.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const AuthError = require('./errors').AuthError;
+
 function node_require(module) {
     return require(module);
 }
@@ -35,14 +37,6 @@ const postHeaders =  {
     'content-type': 'application/json;charset=utf-8',
     'accept': 'application/json'
 };
-
-class AuthError extends Error {
-    constructor(problem) {
-        super(problem.title);
-        Object.assign(this, problem);
-    }
-}
-exports['AuthError'] = AuthError;
 
 function _authenticate(server, json, callback) {
     json.app_id = '';
